requirements.txt
nltk
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
Pillow
opencv-python
sentence-transformers
flair
pycrypto
numpy==1.20